Deal with explicit '' arguments
If a string from the addP39s script gets piped into this, but has a
'blank' argument, that will come as:

  add_full_P39.js Q936913 1707-05-01 1709-02-03 '' Q335636

But the 'replaces' here will be a string of two quote-marks, not an
empty argument, so we need to explicitly check for that.

This seems like a really hacky workaround to something that should
ideally be fixed at a different level, so I'm hoping I can back this out
again later.